# api/openapi.yaml
openapi: 3.0.3
info:
  title: intern-api
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    V1 API contract for Intern at `intern.bekti.com`.

    Authentication modes:

    - browser traffic arrives behind trusted forward-auth and is normalized by the API
    - public clients authenticate with bearer JWT access tokens minted by the device code flow

    Authorization is enforced server-side. Admin-only operations require membership in the configured admin group set, which defaults to `Super-Users`.
servers:
  - url: https://intern.bekti.com
    description: Production
  - url: http://intern-api:8080
    description: Local Compose network
...
paths:
...
  /api/v1/admin/auth/sessions:
    get:
      operationId: listAdminAuthSessions
      tags: [Admin]
      summary: List active auth sessions across users.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 25
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: Active sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/auth/sessions/revoke_all:
    post:
      operationId: revokeAllAdminAuthSessions
      tags: [Admin]
      summary: Revoke all active auth sessions as an administrator.
      description: Revokes all active API/client sessions across all users, including the administrator's own API/client sessions. Browser forward-auth sessions are not affected.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All active API/client sessions revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/auth/sessions/{id}/revoke:
    post:
      operationId: revokeAdminAuthSession
      tags: [Admin]
      summary: Revoke a session as an administrator.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session revoked or already absent.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit_logs:
    get:
      operationId: listAdminAuditLogs
      tags: [Admin]
      summary: List audit logs across the platform.
      security:
        - bearerAuth: []
      parameters:
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: resource_type
          in: query
          required: false
          schema:
            type: string
        - name: resource_id
          in: query
          required: false
          schema:
            type: string
        - name: actor_username
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: Audit log page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/auth/device_codes:
    post:
      operationId: createDeviceCode
      tags: [Auth]
      summary: Create a new device code request for a public client.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCodeCreateRequest'
      responses:
        '201':
          description: Device code request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCode'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/device_codes/{user_code}/approve:
    post:
      operationId: approveDeviceCode
      tags: [Auth]
      summary: Approve a pending device code request for the current authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserCode'
      responses:
        '204':
          description: Authorization approved.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/device_codes/{user_code}/deny:
    post:
      operationId: denyDeviceCode
      tags: [Auth]
      summary: Deny a pending device code request for the current authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserCode'
      responses:
        '204':
          description: Authorization denied.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/tokens:
    post:
      operationId: exchangeDeviceCode
      tags: [Auth]
      summary: Exchange an approved device code for access and refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCodeTokenRequest'
      responses:
        '200':
          description: Tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid, denied, or expired device code request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientAuthError'
        '428':
          description: Device code request still pending approval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientAuthError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/tokens/refresh:
    post:
      operationId: refreshAccessToken
      tags: [Auth]
      summary: Rotate the refresh token and mint a new access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens rotated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/logout:
    post:
      operationId: logoutSession
      tags: [Auth]
      summary: Revoke the presented refresh session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Session revoked or already absent.
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserCode:
      name: user_code
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or credentials invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Caller is authenticated but not authorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    AuthSession:
      type: object
      required:
        - id
        - username
        - client_name
        - created_at
        - expires_at
        - idle_expires_at
        - is_current
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        client_name:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
        idle_expires_at:
          type: string
          format: date-time
        is_current:
          type: boolean
    AuthSessionPagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    AuthSessionPage:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuthSession'
        pagination:
          $ref: '#/components/schemas/AuthSessionPagination'
    AuditLogEntry:
      type: object
      required:
        - id
        - actor_username
        - action
        - resource_type
        - resource_id
        - metadata
        - created_at
      properties:
        id:
          type: string
          format: uuid
        actor_username:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditLogPagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    AuditLogList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        pagination:
          $ref: '#/components/schemas/AuditLogPagination'
...
    DeviceCodeCreateRequest:
      type: object
      properties:
        client_name:
          type: string
          default: internctl
    DeviceCode:
      type: object
      required:
        - device_code
        - user_code
        - verification_uri
        - verification_uri_complete
        - expires_in
        - interval
      properties:
        device_code:
          type: string
        user_code:
          type: string
        verification_uri:
          type: string
          format: uri
        verification_uri_complete:
          type: string
          format: uri
        expires_in:
          type: integer
          format: int32
        interval:
          type: integer
          format: int32
    DeviceCodeTokenRequest:
      type: object
      required: [device_code]
      properties:
        device_code:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in_seconds, refresh_token]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in_seconds:
          type: integer
          format: int32
        refresh_token:
          type: string
    ClientAuthError:
      type: object
      required: [error, error_description]
      properties:
        error:
          type: string
          enum:
            - authorization_pending
            - invalid_request
            - access_denied
            - expired_token
            - slow_down
        error_description:
          type: string